move provider settings
